feat: useCase를 구현하는 applicationService / jpa 관계 매핑에 따른 queryDsl 역할 최소화

+++ src/main/kotlin/kr/crispyblog/moneynote/domain/money/repositories/PeriodRepository.kt
@@ -8,5 +8,5 @@ import java.util.*
 
 interface PeriodRepository : JpaRepository<Period, Long> {
 
-    fun findByUserIdAndStartDateGreaterThanEqualAndEndDateLessThanEqual(userId: Int, startDate: LocalDate?, endDate: LocalDate?): Optional<List<Period>>
+    fun findByUserIdAndStartDateGreaterThanEqualAndEndDateLessThanEqual(userId: Long, startDate: LocalDate?, endDate: LocalDate?): Optional<List<Period>>
 }
